fix opcodes arguments
