+++ src/opcodes.c
@@ -1,40 +1,121 @@
 #include <stdint.h>
+#include <stdio.h>
+
+#include "chip8.h"
 
 //******************************************************************************
-//*                     OcpuData->pcODES FUNCTIONS DEFINITIONS                          *
+//*                     OcpuData->pcODES FUNCTIONS DEFINITIONS                 *
 //******************************************************************************
+
+// call our opcodes according to the function pointers
+void msbis0(uint16_t opcode)
+{
+    if (opcode)
+    {
+        uint16_t index = offset4(opcode);
+        (*zeroop[index]) (opcode);
+    }
+}
+
+void msbis8(uint16_t opcode)
+{
+    uint16_t index = offset4(opcode);
+    (*eightop[index]) (opcode);
+}
+
+void msbise(uint16_t opcode) 
+{
+    uint16_t index = offset3(opcode);
+    (*e_op[index]) (opcode);
+}
+
+void msbisf(uint16_t opcode)
+{
+    uint16_t index = offset4(opcode);
+    if (index == 0x5) {
+        index = offset3(opcode);
+    }
+
+    (*special[index]) (opcode);
+}
+
+
+// Handle opcodes starting with 0x0
+void (*zeroop[]) (uint16_t opcode) =
+{
+    cls, cpuNULL, cpuNULL, cpuNULL, cpuNULL, 
+    cpuNULL, cpuNULL, cpuNULL, cpuNULL, cpuNULL, 
+    cpuNULL, cpuNULL, cpuNULL, cpuNULL, ret
+};
+
+// Handle opcodes starting with 0x8
+void (*eightop[]) (uint16_t opcode) = 
+{
+    setvxtovy, vxorvy, vxandvy, vxxorvy, 
+    vxaddvy, vxsubvy, lsb_vx_in_vf_r, 
+    vysubvx, cpuNULL, cpuNULL, cpuNULL, 
+    cpuNULL, cpuNULL, cpuNULL, 
+    svflsl
+};
+
+// Handle opcodes starting with 0xE
+void (*e_op[]) (uint16_t opcode) = 
+{
+    cpuNULL, cpuNULL, cpuNULL, cpuNULL, 
+    cpuNULL, cpuNULL, cpuNULL, cpuNULL, 
+    cpuNULL, skipifdown, skipnotdown
+};
+
+// Handle opcodes starting with 0xF
+void (*special[]) (uint16_t opcode) =
+{
+    cpuNULL, set_dt, cpuNULL, set_BCD, cpuNULL, 
+    reg_dump, reg_load, vx_to_dt, set_st, load_char_addr, 
+    vx_to_key, cpuNULL, cpuNULL, cpuNULL, iaddvx, cpuNULL
+
+};
+
+// Call other arrays of functions 
+void (*generalop[16]) (uint16_t opcode) =
+{
+    msbis0, jump, call, se, sne, 
+    svxevy, setvx, addvx, msbis8, next_if_vx_not_vy, 
+    itoa, jmpaddv0, vxandrand, draw, msbise, 
+    msbisf
+};
+
 
 // fuctions for the general function pointers array
 
 void msbis0(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    if (ocpuData->pcode)
+    if (opcode)
     {
-        uint16_t index = offset4(ocpuData->pcode);
-        (*zeroop[index]) (ocpuData->pcode);
+        uint16_t index = offset4(opcode);
+        (*zeroop[index]) (opcode);
     }
 }
 
 void msbis8(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint16_t index = offset4(ocpuData->pcode);
-    (*eightop[index]) (ocpuData->pcode);
+    uint16_t index = offset4(opcode);
+    (*eightop[index]) (opcode);
 }
 
 void msbise(uint16_t opcode, cpu *cpuData, MemMaps *mem) 
 {
-    uint16_t index = offset3(ocpuData->pcode);
-    (*e_op[index]) (ocpuData->pcode);
+    uint16_t index = offset3(opcode);
+    (*e_op[index]) (opcode);
 }
 
 void msbisf(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint16_t index = offset4(ocpuData->pcode);
+    uint16_t index = offset4(opcode);
     if (index == 0x5) {
-        index = offset3(ocpuData->pcode);
+        index = offset3(opcode);
     }
 
-    (*special[index]) (ocpuData->pcode);
+    (*special[index]) (opcode);
 
 }
 
@@ -44,16 +125,16 @@ void setvx(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
     // set register specified in msb - 4 bits to 
     // the lsb
-    uint8_t vx = offset2(ocpuData->pcode);
-    uint8_t nn = ocpuData->pcode & 0x00ff;
+    uint8_t vx = offset2(opcode);
+    uint8_t nn = opcode & 0x00ff;
 
     cpuData->regs[vx] = nn;
 }
 
 void setvxtovy(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = offset2(ocpuData->pcode);
-    uint8_t vy = offset3(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
+    uint8_t vy = offset3(opcode);
 
     cpuData->regs[vx] = cpuData->regs[vy];
 }
@@ -61,16 +142,16 @@ void setvxtovy(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 void addvx(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {   
     uint8_t nn = ocpuData->pcode & 0x00ff;
-    uint8_t vx = offset2(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
 
     cpuData->regs[vx] = cpuData->regs[vx] + nn;
 }
 
 void vxaddvy(uint16_t opcode, cpu *cpuData, MemMaps *mem) 
 {
-    //printf("running vxaddvy\n", ocpuData->pcode);
-    uint8_t vx = offset2(ocpuData->pcode);
-    uint8_t vy = offset3(ocpuData->pcode);
+    //printf("running vxaddvy\n", opcode);
+    uint8_t vx = offset2(opcode);
+    uint8_t vy = offset3(opcode);
     uint16_t result;
 
     result = cpuData->regs[vx] + cpuData->regs[vy];
@@ -83,9 +164,9 @@ void vxaddvy(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 
 void vxsubvy(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    //printf("running vxsubvy %X\n", ocpuData->pcode);
-    uint8_t vx = offset2(ocpuData->pcode);
-    uint8_t vy = offset3(ocpuData->pcode);
+    //printf("running vxsubvy %X\n", opcode);
+    uint8_t vx = offset2(opcode);
+    uint8_t vy = offset3(opcode);
 
     cpuData->regs[vx] = cpuData->regs[vx] - cpuData->regs[vy];
     uint8_t borrow = cpuData->regs[vy] > cpuData->regs[vx];
@@ -95,8 +176,8 @@ void vxsubvy(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 
 void vysubvx(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {  
-    uint8_t vx = offset2(ocpuData->pcode);
-    uint8_t vy = offset3(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
+    uint8_t vy = offset3(opcode);
 
     cpuData->regs[vx] = cpuData->regs[vy] - cpuData->regs[vx];
     uint8_t borrow = cpuData->regs[vx] < cpuData->regs[vy];
@@ -106,32 +187,32 @@ void vysubvx(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 
 void vxorvy(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = offset2(ocpuData->pcode);
-    uint8_t vy = offset3(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
+    uint8_t vy = offset3(opcode);
 
     cpuData->regs[vx] = cpuData->regs[vx] | cpuData->regs[vy];
 }
 
 void vxandvy(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = offset2(ocpuData->pcode);
-    uint8_t vy = offset3(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
+    uint8_t vy = offset3(opcode);
 
     cpuData->regs[vx] = cpuData->regs[vx] & cpuData->regs[vy];
 }
 
 void vxxorvy(uint16_t opcode, cpu *cpuData, MemMaps *mem) 
 {
-    uint8_t vx = offset2(ocpuData->pcode);
-    uint8_t vy = offset3(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
+    uint8_t vy = offset3(opcode);
 
     cpuData->regs[vx] = cpuData->regs[vx] ^ cpuData->regs[vy];
 }
 
 void lsb_vx_in_vf_r(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = offset2(ocpuData->pcode);
-    uint8_t vy = offset3(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
+    uint8_t vy = offset3(opcode);
 
     reg[0xf] = cpuData->regs[vx] & 0x01;
     cpuData->regs[vx] = cpuData->regs[vx] >> 1;
@@ -141,8 +222,8 @@ void lsb_vx_in_vf_r(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 
 void svflsl(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = offset2(ocpuData->pcode);
-    uint8_t vy = offset3(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
+    uint8_t vy = offset3(opcode);
 
     // store msb of vx in vf
     reg[0xf] = (cpuData->regs[vx] & 0x80) >> 7;
@@ -153,7 +234,7 @@ void svflsl(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 
 void vxandrand(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = offset2(ocpuData->pcode);    // register index
+    uint8_t vx = offset2(opcode);    // register index
 
     uint8_t mask = ocpuData->pcode & 0x00ff;  // mask value
     uint8_t rand = randnum();        // random number
@@ -193,7 +274,7 @@ void call(uint16_t opcode, cpu *cpuData, MemMaps *mem)
     ++SP;
     stack[SP] = cpuData->pc;
 
-    cpuData->pc = (0x0FFF & ocpuData->pcode) - 2;
+    cpuData->pc = (0x0FFF & opcode) - 2;
 }
 
 void ret(uint16_t opcode, cpu *cpuData, MemMaps *mem)
@@ -209,7 +290,7 @@ void ret(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 
 void se(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {   
-    uint8_t vx = offset2(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
 
     uint8_t nn = 0x00FF & ocpuData->pcode;
 
@@ -221,8 +302,8 @@ void se(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 
 void svxevy(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = offset2(ocpuData->pcode);
-    uint8_t vy = offset3(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
+    uint8_t vy = offset3(opcode);
 
 
     if (cpuData->regs[vx] == cpuData->regs[vy]) 
@@ -233,7 +314,7 @@ void svxevy(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 
 void sne(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = offset2(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
     uint8_t nn = ocpuData->pcode & 0x00FF;
 
     if (cpuData->regs[vx] != nn) 
@@ -244,8 +325,8 @@ void sne(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 
 void next_if_vx_not_vy(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 { 
-    uint8_t vx = offset2(ocpuData->pcode);
-    uint8_t vy = offset3(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
+    uint8_t vy = offset3(opcode);
 
     if (cpuData->regs[vx] != cpuData->regs[vy])
     {
@@ -257,21 +338,21 @@ void next_if_vx_not_vy(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 
 void set_dt(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = offset2(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
 
     DT = cpuData->regs[vx];
 }
 
 void vx_to_dt(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = offset2(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
 
     cpuData->regs[vx] = DT;
 }
 
 void set_st(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = offset2(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
 
     ST = cpuData->regs[vx];
 }
@@ -281,13 +362,13 @@ void set_st(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 
 void vx_to_key(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = offset2(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
     cpuData->regs[vx] = waitkey();
 }
 
 void skipifdown(uint16_t opcode, cpu *cpuData, MemMaps *mem) 
 {
-    uint8_t vx = reg[offset2(ocpuData->pcode)]; // value stored in register vx
+    uint8_t vx = reg[offset2(opcode)]; // value stored in register vx
     uint8_t is_pressed = keys[vx];     // 1 if key in vx is pressed, 0 otherwise
 
     // skip if key is pressed
@@ -299,7 +380,7 @@ void skipifdown(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 
 void skipnotdown(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = reg[offset2(ocpuData->pcode)]; // value stored in register vx
+    uint8_t vx = reg[offset2(opcode)]; // value stored in register vx
     uint8_t is_pressed = keys[vx];     // 1 if key in vx is pressed, 0 otherwise
 
     // skip instruction if key is not pressed
@@ -320,7 +401,7 @@ void itoa(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 
 void iaddvx(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = offset2(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
     I = I + cpuData->regs[vx];
 
 }
@@ -330,9 +411,9 @@ void iaddvx(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 void draw(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
     // where to draw and how many bytes
-    uint16_t x = reg[offset2(ocpuData->pcode)];
-    uint16_t y = reg[offset3(ocpuData->pcode)];
-    uint16_t height = offset4(ocpuData->pcode);
+    uint16_t x = reg[offset2(opcode)];
+    uint16_t y = reg[offset3(opcode)];
+    uint16_t height = offset4(opcode);
 
     // byte of data containing the pixel to be written
     uint8_t pixel;
@@ -381,7 +462,7 @@ void cls(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 void load_char_addr(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
     // this is the hexadecimal character
-    uint8_t hex = reg[offset2(ocpuData->pcode)];
+    uint8_t hex = reg[offset2(opcode)];
 
     // set I to the starting address of the data that composes the
     // hexadecimal digit in memory
@@ -396,7 +477,7 @@ void set_BCD(uint16_t opcode, cpu *cpuData, MemMaps *mem)
     int i;
 
     uint8_t digits[3];
-    uint8_t number = reg[offset2(ocpuData->pcode)]; 
+    uint8_t number = reg[offset2(opcode)]; 
 
     // store separate digits into the digits array
     for (i = 3; number > 0;)
@@ -413,7 +494,7 @@ void set_BCD(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 
 void reg_dump(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = offset2(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
     uint8_t index;
 
     for (index = 0x0; index <= vx; ++index)
@@ -425,7 +506,7 @@ void reg_dump(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 
 void reg_load(uint16_t opcode, cpu *cpuData, MemMaps *mem)
 {
-    uint8_t vx = offset2(ocpuData->pcode);
+    uint8_t vx = offset2(opcode);
     uint8_t index;
 
     for (index = 0x0; index <= vx; ++index)
@@ -433,4 +514,9 @@ void reg_load(uint16_t opcode, cpu *cpuData, MemMaps *mem)
         reg[index] = ram[I+index];
     }
     //I = I + vx + 1;
+}
+
+void cpuNULL(uint16_t opcode, cpu *cpuData, MemMaps *mem)
+{
+    fprintf(stderr, "[WARNING] Unknown opcode %#X at %#X\n", opcode, PC);
 }
